+++ src/main/java/com/chemax/project/service/EquipmentTypeService.java
@@ -1,9 +1,9 @@
 package com.chemax.project.service;
 
-import com.chemax.project.dto.EquipmentDTO;
 import com.chemax.project.dto.EquipmentTypeDTO;
 import com.chemax.project.entities.EquipmentTypeEntity;
 import com.chemax.project.exceptions.EntityNotFoundException;
+import com.chemax.project.exceptions.NeedToMoveEntityException;
 import com.chemax.project.repository.EquipmentTypeEntityRepository;
 import com.chemax.project.request.EquipmentTypeRequest;
 import org.springframework.stereotype.Service;
@@ -20,7 +20,8 @@ public class EquipmentTypeService {
     private final EquipmentTypeEntityRepository equipmentTypeRepository;
     private final EquipmentService equipmentService;
 
-    public EquipmentTypeService(EquipmentTypeEntityRepository equipmentTypeRepository, EquipmentService equipmentService) {
+    public EquipmentTypeService(EquipmentTypeEntityRepository equipmentTypeRepository,
+                                EquipmentService equipmentService) {
         this.equipmentTypeRepository = equipmentTypeRepository;
         this.equipmentService = equipmentService;
     }
@@ -57,12 +58,18 @@ public class EquipmentTypeService {
     }
 
     public void deleteEquipmentTypeEntity (Integer id) {
-        equipmentTypeRepository.delete(getEquipmentTypeEntity(id));
+        EquipmentTypeEntity equipmentTypeEntityToDelete = equipmentTypeRepository.getReferenceById(id);
+        if (equipmentTypeEntityToDelete.getEquipmentList().isEmpty()) {
+            equipmentTypeRepository.delete(getEquipmentTypeEntity(id));
+        } else {
+            throw new NeedToMoveEntityException();
+        }
     }
 
-    public void updateEquipmentTypeEntity (EquipmentTypeRequest request, Integer id) {
+    public void updateEquipmentTypeEntity (EquipmentTypeDTO equipmentTypeDTO, Integer id) {
         EquipmentTypeEntity equipmentTypeEntity = equipmentTypeRepository.getReferenceById(id);
-        equipmentTypeEntity.setMachineType(Optional.ofNullable(request.getMachineType()).orElse(equipmentTypeEntity.getMachineType()));
+        equipmentTypeEntity.setMachineType(Optional.ofNullable(equipmentTypeDTO.getMachineType())
+                .orElse(equipmentTypeEntity.getMachineType()));
         equipmentTypeRepository.save(equipmentTypeEntity);
     }
 
@@ -77,7 +84,8 @@ public class EquipmentTypeService {
         equipmentTypeDTO.setId(equipmentTypeEntityToConvert.getId());
         equipmentTypeDTO.setMachineType(equipmentTypeEntityToConvert.getMachineType());
         equipmentTypeDTO.setEquipmentDTOList(equipmentService.getAllEquipmentDTOs().stream()
-                .filter(equipmentDTO -> Objects.equals(equipmentDTO.getMachineTypeId(), equipmentTypeEntityToConvert.getId())).collect(Collectors.toList()));
+                .filter(equipmentDTO -> Objects.equals(equipmentDTO.getMachineTypeId(), equipmentTypeEntityToConvert
+                        .getId())).collect(Collectors.toList()));
         return equipmentTypeDTO;
     }
 }
